# swagger.yaml
openapi: 3.0.0
info:
  title: Reactors API Documentation
  version: 1.0.0
  description: API documentation for the Reactors Book Tracker project
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HttpOnly cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 691ce9f36a70ea178440d121
        name:
          type: string
          example: Jawad Taki
        email:
          type: string
          format: email
          example: jawadtaki0@gmail.com
        profilePic:
          type: string
          format: uri
          nullable: true
        authProvider:
          type: string
          example: google
    AuthUserResponse:
      type: object
      properties:
        message:
          type: string
          example: User authenticated
        user:
          $ref: "#/components/schemas/User"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong
        error:
          type: string
          nullable: true
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john@example.com
        password:
          type: string
          example: StrongP@ssw0rd
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: john@example.com
        password:
          type: string
          example: StrongP@ssw0rd
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
    OtpVerifyRequest:
      type: object
      required:
        - userId
        - otp
      properties:
        userId:
          type: string
          example: 691ce9f36a70ea178440d121
        otp:
          type: string
          example: "233852"
    OtpVerifyResponse:
      type: object
      properties:
        message:
          type: string
          example: OTP verified successfully
        user:
          $ref: "#/components/schemas/User"
    ResendOtpRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          example: 691ce9f36a70ea178440d121
...
paths:
  /api/auth/google:
    get:
      tags: [Auth]
      summary: Start Google OAuth login
      description: Redirects the user to Google for OAuth login.
      responses:
        "302":
          description: Redirect to Google OAuth consent screen
        "500":
          description: Google OAuth Error
  /api/auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      description: Handles the callback from Google, creates the user (if new), and sends MFA OTP to email.
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: Authorization code returned by Google
      responses:
        "200":
          description: MFA required, OTP sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: MFA required
                  email:
                    type: string
                    example: jawadtaki0@gmail.com
                  userId:
                    type: string
        "400":
          description: Missing or invalid code
        "500":
          description: Google OAuth Error
  /api/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify MFA OTP
      description: Verifies the OTP sent to the user's email, resets OTP fields, and issues a JWT cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified and user logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid OTP, user not found, or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many failed attempts, user locked temporarily
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/resend-otp:
    post:
      tags: [Auth]
      summary: Resend MFA OTP
      description: Generates and emails a new OTP code to the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendOtpRequest"
      responses:
        "200":
          description: New OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: A new OTP has been sent to your email.
        "400":
          description: Missing userId or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (email/password)
      description: Registers a new user with name, email, and password. Sets a JWT cookie on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation errors (name, email, or password) or email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Logs in a user using email/password and sets a JWT in an HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
  /api/auth/logout:
    post:
      security:
        - cookieAuth: []
      tags: [Auth]
      summary: Logout the current user
      description: Clears the JWT cookie.
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
  /api/auth/me:
    get:
      security:
        - cookieAuth: []
      tags: [Auth]
      summary: Get the current logged-in user
      description: Returns the authenticated user based on the JWT in the cookie.
      responses:
        "200":
          description: User authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: No valid token / not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...